adjust the class structure and fix bugs

+++ Example/DLCMobilePlayer/DLCViewController.m
@@ -39,7 +39,7 @@
 
 #pragma mark - Event
 - (IBAction)videoSnapshotAction:(UIButton *)sender {
-    NSString *key = [NSString stringWithFormat:@"%ld", (NSInteger)([[NSDate date] timeIntervalSince1970]*1000)];
+    NSString *key = [NSString stringWithFormat:@"%ld", (long)([[NSDate date] timeIntervalSince1970]*1000)];
     
     self.snapshotImageView.image = [self.videoView takeVideoSnapshotImageForKey:key];
     NSLog(@"video snapshot path: %@", [DLCSubVideoView videoSnapshotPathForKey:key]);
